+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * raised minimum required PHP version to 7.3.0
 * added more type hints
+* `stubbles\db\config\DatabaseConfiguration::getInitialQuery()` now returns an empty string if no initial query configured
 
 ## 8.0.0 (2016-07-31)